+++ pkg/core/types.go
@@ -1,0 +1,10 @@
+package core
+
+type Storage[T any] interface {
+	Append(T) error
+	GetRecords() ([]T, error)
+}
+
+type ValueRequester interface {
+	GetValue() (float64, error)
+}
